requirements.txt
pandas 
numpy
matplotlib
seaborn
mysql-connector-python
pymysql
python-dotenv
scikit-learn